Assignment 3-4/assignment-3-muhammed786fiyas/requirements.txt
# Core
numpy
pandas
Pillow
opencv-python
PyYAML

# Machine Learning
scikit-learn
scikit-image
scipy
tensorflow
keras

# DVC
dvc

# Visualization
matplotlib